+++ packages/v4/src/content/design-guidelines/styles/typography/typography.js
@@ -7,11 +7,12 @@ export const PfStyledText = ({ children, variableName, fontWeight, lineHeight, f
   </div>
 );
 
-export const TitleLevel = ({ asGrid, styleProps, children, title = '', note = '' }) => {
+export const TitleLevel = ({ asGrid, styleProps, children, title = '', note = '', className = '' }) => {
   return asGrid ? (
     <TypographyGrid 
       title={title} 
       note={note}
+      className={className}
       {...styleProps}
     />
   ) : (
@@ -25,55 +26,70 @@ export const styleProps = {
     fontWeightText: "400 (medium)",
     variableName: "--pf-global--FontSize--2xl",
     fontSize: "24px",
-    lineHeight: "1.3"
+    lineHeight: "1.3",
+    fontFamily: "RedHatDisplay"
   },
   second: {
     fontWeight: "400",
     fontWeightText: "400 (medium)",
     fontSize: "20px",
     variableName: "--pf-global--FontSize--xl",
-    lineHeight: "1.3"
+    lineHeight: "1.3",
+    fontFamily: "RedHatDisplay"
   },
   third: {
     fontWeight: "400",
     fontWeightText: "400 (medium)",
     fontSize: "18px",
     variableName: "--pf-global--FontSize--lg",
-    lineHeight: "1.5"
+    lineHeight: "1.5",
+    fontFamily: "RedHatDisplay"
   },
   fourth: {
-    fontWeight: "700",
-    fontWeightText: "700 (bold)",
+    fontWeight: "400",
+    fontWeightText: "400 (medium)",
     fontSize: "16px",
     variableName: "--pf-global--FontSize--md",
-    lineHeight: "1.5"
+    lineHeight: "1.5",
+    fontFamily: "RedHatDisplay"
   },
   body: {
     fontWeight: "400",
     fontWeightText: "400 (regular)*",
     fontSize: "16px",
     variableName: "--pf-global--FontSize--md",
-    lineHeight: "1.5"
+    lineHeight: "1.5",
+    fontFamily: "RedHatText"
   },
   small: {
     fontWeight: "400",
     fontWeightText: "400 (regular)",
     fontSize: "14px",
     variableName: "--pf-global--FontSize--sm",
-    lineHeight: "1.5"
+    lineHeight: "1.5",
+    fontFamily: "RedHatText"
   },
   tiny: {
     fontWeight: "400",
     fontWeightText: "400 (regular)",
     fontSize: "12px",
     variableName: "--pf-global--FontSize--xs",
-    lineHeight: "1.5"
+    lineHeight: "1.5",
+    fontFamily: "RedHatText"
+  },
+  code: {
+    fontWeight: "400",
+    fontWeightText: "400 (regular)",
+    fontSize: "16px",
+    variableName: "--pf-global--FontSize--md",
+    lineHeight: "1.5",
+    fontFamily: "Liberation Mono"
   }
 }
 
-export const TypographyGrid = ({title, note, symbol, fontWeight, fontWeightText, fontSize, variableName, lineHeight, fontFamily}) => (
+export const TypographyGrid = ({title, note, symbol, fontWeight, fontWeightText, fontSize, variableName, lineHeight, fontFamily, className}) => (
   <React.Fragment>
-    <h3 className="ws-title">{title} {symbol && <span className="ws-typography-gridTitleSymbol">{symbol}</span>}</h3>
+    <h3 className={className ? `ws-title ${className}` : 'ws-title'}>{title} {symbol && <span className="ws-typography-gridTitleSymbol">{symbol}</span>}</h3>
     <PfStyledText fontFamily={fontFamily} fontWeight={fontWeight} variableName={variableName} lineHeight={lineHeight}>
       Design is where science and art break even.
     </PfStyledText>
@@ -81,6 +97,10 @@ export const TypographyGrid = ({title, note, symbol, fontWeight, fontWeightText,
       <GridItem span={12}>
         <table className="pf-c-table pf-m-compact ws-typography-tableTypography" aria-label="typography usage guidelines breakout">
           <tbody>
+            <tr>	
+              <td className="pf-u-pr-sm">Font family:</td>	
+              <td>{fontFamily}</td>	
+            </tr>
             <tr>
               <td>Font weight:</td>
               <td>{fontWeightText}</td>
@@ -127,7 +147,7 @@ export const Spacer = ({color, size, description, showBorder}) => {
     backgroundColor = '#1b8793';
   }
   return (
-    <div style={{display: "flex", padding: description ? "15px" : "0px"}}>
+    <div className="pf-l-flex pf-m-space-items-md pf-m-align-items-center">
       {description && <div>{description}</div>}
       <div className={`
         ws-typography-spacerBox-size-${size}